Missing explicit type cast, again

+++ Lilu/Sources/kern_patcher.cpp
@@ -451,7 +451,7 @@ kern_return_t KernelPatcher::onVmMapEnterMemObjectControl(
 			if (that->kcMachInfos[kcType] != nullptr) {
 				uint8_t *patchedKC = that->kcMachInfos[kcType]->file_buf;
 				SYSLOG("patcher", "onVmMapEnterMemObjectControl: Copying %sKC range %llX ~ %llX", kcType, offset, offset + initial_size);
-				memcpy((void*)*address, patchedKC + offset, initial_size);
+				memcpy((void*)*address, patchedKC + offset, (size_t)initial_size);
 			}
 		}
 	}
